completed exercise 4.14 - 4.22

+++ part4/bloglist/controllers/blogs.js
@@ -1,24 +1,37 @@
 const blogsRouter = require('express').Router()
 const Blog = require('../models/blog')
+const User = require('../models/user')
+const jwt = require('jsonwebtoken')
 
-blogsRouter.get('/', async (request, response,next) => {
-  try {
-    const blogs = await Blog.find({})
-    response.json(blogs)
-  }
-  catch(exception)
-  {
-    next(exception)
-  }
+
+blogsRouter.get('/', async (request, response) => {
+
+  const blogs = await Blog.find({}).populate('user', { username: 1, name: 1 })
+  response.json(blogs)
 })
 
 blogsRouter.post('/', async (request, response) => {
-  const blog = new Blog(request.body)
+  const body = request.body
+
+  const user = request.user
+
+  const blog = new Blog({
+    title: body.title,
+    author: body.author,
+    url: body.url,
+    likes: body.likes,
+    user: user.id
+  })
+
+
   const createdBlog = await blog.save()
+  user.blogs = user.blogs.concat(createdBlog._id)
+  await user.save()
+
   response.status(201).json(createdBlog)
 })
 
-blogsRouter.get('/:id',async (request,response,) => {
+blogsRouter.get('/:id', async (request, response,) => {
   const blog = await Blog.findById(request.params.id)
 
   if (blog) {
@@ -28,16 +41,24 @@ blogsRouter.get('/:id',async (request,response,) => {
   }
 })
 
-blogsRouter.delete('/:id',async (request,response) => {
-  await Blog.findByIdAndRemove(request.params.id)
-  response.status(204).end()
+blogsRouter.delete('/:id', async (request, response) => {
+  
+  const user = request.user
+  const blog = await Blog.findById(request.params.id)
+
+  if (blog.user.toString() === user.id) {
+    await Blog.findByIdAndRemove(request.params.id)
+    response.status(204).end()
+  }else {
+    return response.status(401).json({ error: 'invalid user' })
+  }
 })
 
-blogsRouter.put('/:id', async (request,response) => {
+blogsRouter.put('/:id', async (request, response) => {
   const blog = request.body
-  const updatedBlog = await Blog.findByIdAndUpdate(request.params.id,blog,{ new: true })
+  const updatedBlog = await Blog.findByIdAndUpdate(request.params.id, blog, { new: true })
 
-  if(updatedBlog)
+  if (updatedBlog)
     response.status(200).json(updatedBlog.toJSON())
   else
     response.status(404).end()
